requirements.txt
alembic==1.13.2
bidict==0.23.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SocketIO==5.3.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.0
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
numpy==2.1.1
pandas==2.2.2
PyJWT==2.9.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.4
pytz==2024.2
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.34
typing_extensions==4.12.2
tzdata==2024.1
Werkzeug==3.0.4
wsproto==1.2.0
